requirements/base.txt
# Base Python Dependencies
# Common packages used across all services

# Core
numpy==1.24.3
scipy==1.11.4
pandas==2.0.3
pillow==10.2.0
opencv-python==4.9.0.80

# Data handling
pyyaml==6.0.1
python-dotenv==1.0.0
tqdm==4.66.1
requests==2.31.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# 3D processing
trimesh==4.0.10
pygltflib==1.16.1
Pillow==10.2.0

# Utilities
click==8.1.7
colorama==0.4.6
rich==13.7.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.12.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Logging
loguru==0.7.2
